+++ internal/service/service_test.go
@@ -33,7 +33,7 @@ func TestHandler(t *testing.T) {
 	for name, tc := range tests {
 		t.Run(name, func(t *testing.T) {
 			// setup
-			s := service.OPDS{"testdata"}
+			s := service.OPDS{"testdata", false}
 			w := httptest.NewRecorder()
 			req := httptest.NewRequest(http.MethodGet, tc.input, nil)
 			service.TimeNow = func() time.Time {
@@ -68,22 +68,22 @@ var feed = `<?xml version="1.0" encoding="UTF-8"?>
           <title>emptyFolder</title>
           <id>/emptyFolder</id>
           <link rel="subsection" href="/emptyFolder" type="application/atom+xml;profile=opds-catalog;kind=acquisition" title="emptyFolder"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
       <entry>
           <title>mybook</title>
           <id>/mybook</id>
           <link rel="subsection" href="/mybook" type="application/atom+xml;profile=opds-catalog;kind=acquisition" title="mybook"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
       <entry>
           <title>new folder</title>
           <id>/new folder</id>
           <link rel="subsection" href="/new%20folder" type="application/atom+xml;profile=opds-catalog;kind=acquisition" title="new folder"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
   </feed>`
 
@@ -97,35 +97,35 @@ var acquisitionFeed = `<?xml version="1.0" encoding="UTF-8"?>
           <title>mybook copy.epub</title>
           <id>/mybookmybook copy.epub</id>
           <link rel="http://opds-spec.org/acquisition" href="/mybook/mybook%20copy.epub" type="application/epub+zip" title="mybook copy.epub"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
       <entry>
           <title>mybook copy.txt</title>
           <id>/mybookmybook copy.txt</id>
           <link rel="http://opds-spec.org/acquisition" href="/mybook/mybook%20copy.txt" type="text/plain; charset=utf-8" title="mybook copy.txt"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
       <entry>
           <title>mybook.epub</title>
           <id>/mybookmybook.epub</id>
           <link rel="http://opds-spec.org/acquisition" href="/mybook/mybook.epub" type="application/epub+zip" title="mybook.epub"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
       <entry>
           <title>mybook.pdf</title>
           <id>/mybookmybook.pdf</id>
           <link rel="http://opds-spec.org/acquisition" href="/mybook/mybook.pdf" type="application/pdf" title="mybook.pdf"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
       <entry>
           <title>mybook.txt</title>
           <id>/mybookmybook.txt</id>
           <link rel="http://opds-spec.org/acquisition" href="/mybook/mybook.txt" type="text/plain; charset=utf-8" title="mybook.txt"></link>
-          <published>2020-05-25T00:00:00+00:00</published>
-          <updated>2020-05-25T00:00:00+00:00</updated>
+          <published></published>
+          <updated></updated>
       </entry>
   </feed>`